Refactor delegator_pool models
