+++ rust/processor/src/db/common/models/mod.rs
@@ -1,3 +1,4 @@
 pub mod ans_models;
 pub mod default_models;
 pub mod event_models;
+pub mod stake_models;
