+++ rust/processor/src/db/common/models/stake_models/mod.rs
@@ -1,0 +1,2 @@
+pub mod delegator_pools;
+pub mod stake_utils;
