+++ rust/processor/src/db/postgres/models/stake_models/delegator_activities.rs
@@ -3,7 +3,7 @@
 // This is required because a diesel macro makes clippy sad
 #![allow(clippy::extra_unused_lifetimes)]
 
-use super::stake_utils::StakeEvent;
+use crate::db::common::models::stake_models::stake_utils::StakeEvent;
 use crate::{
     schema::delegated_staking_activities,
     utils::{
