+++ rust/processor/src/db/postgres/models/stake_models/mod.rs
@@ -6,5 +6,4 @@ pub mod delegator_activities;
 pub mod delegator_balances;
 pub mod delegator_pools;
 pub mod proposal_votes;
-pub mod stake_utils;
 pub mod staking_pool_voter;
