+++ rust/processor/src/db/postgres/models/stake_models/proposal_votes.rs
@@ -4,7 +4,7 @@
 // This is required because a diesel macro makes clippy sad
 #![allow(clippy::extra_unused_lifetimes)]
 
-use super::stake_utils::StakeEvent;
+use crate::db::common::models::stake_models::stake_utils::StakeEvent;
 use crate::{
     schema::proposal_votes,
     utils::{
